Rename Graph code
